requirements.txt
streamlit==1.39.0
torch==2.7.1
torchvision==0.22.1
pillow==11.2.1
numpy==2.1.3
opencv-python==4.11.0.86
pandas>=2.0.0
